Extract match_block definition and give it a name
Previously, match blocks would be expressed inline in the definition of 
match expressions. Even though this worked correctly in terms of 
parsing, it was preventing Atom from correctly identifying match blocks 
as foldable.

This commit extracts the match block definition and gives it a name. 
Other than fixing the issue in Atom illustrated above, it also increases 
consistency with other kinds of expressions such as `if`/`else`.